+++ tutorial/src/main/scala-0.22/tutorial/examples01To07/Example01ForComprehensions.scala
@@ -1,4 +1,4 @@
-package tutorial.examples
+package tutorial.examples01To07
 
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext
