+++ tutorial/src/main/scala-0.22/tutorial/examples01To07/Example06FunctionsAreValues.scala
@@ -1,4 +1,4 @@
-package tutorial.examples
+package tutorial.examples01To07
 
 import scala.util.chaining._
 import scala.language.implicitConversions
