+++ tutorial/src/main/scala-0.22/tutorial/examples01To07/Example07Function1Monad.scala
@@ -1,4 +1,4 @@
-package tutorial.examples
+package tutorial.examples01To07
 
 import scala.util.chaining._
 import scala.language.implicitConversions
