+++ tutorial/src/main/scala-2.13/tutorial/examples09/Example09aHelloWithFunction0.scala
